backend/requirements-render.txt
# Render-optimized requirements for NourishAI
# Minimal dependencies to avoid build issues

# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
pymongo==4.6.0
motor==3.3.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP requests and APIs
httpx==0.25.2
requests==2.31.0

# Google Gemini AI integration
google-generativeai==0.3.2

# Data validation
pydantic==2.5.1
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Logging
loguru==0.7.2

# Basic data processing (only if needed)
# numpy==1.24.3
# pandas==2.0.3
